Stop testing Koa on Node.js v6.
It was going to come to this sooner or later, since Node.js v6 is no longer
supported by the Node.js Foundation.  In this case, I'm adding this
exception because after bringing #3229 (2dd0592e), the `koa-bodyparser`
package was updated to a new major version which, itself, dropped Node.js 6
support.

That update to `koa-bodyparser`, which fixes an incorrect/malformed
`Content-length` header calculation is — I think — important enough that we
should make sure it's included in Apollo Server, which currently drives the
underlying version of Koa for all users because of its close coupling with
Koa itself (via the `apollo-server-koa` package).

This micro-framework-management will no longer be a concern with
Apollo Server, particularly because of the introduction of a transport
abstraction, which I've proposed in #3184.

Ref: https://github.com/apollographql/apollo-server/issues/3184