Alternative take for #3300.
While #3300 does correctly make the adjustments that are necessary when
someone's getting started experience involves a deployed version of the
application (for example, if you're deploying your Getting Started
application to Zeit's Now), the additional parameters are not otherwise
necessary.  In fact, introducing them may inadvertently cause users to
enable these, when we intentionally take the step to disable it.

Therefore, I'm suggesting this alternative edit to the Getting Started
guide.