Thread the `TResponse` type through the v2 request pipeline.
Previously, the `GraphQLResponse` was inclusive of the `http` property, even
though it is most certainly not a GraphQL-specific property and is instead a
concern of the transport.

While this work may never see the light of day with the approach we're
taking, I thought I'd commit it for posterity's sake, since I did actually
go through the pain of sorting most of it out.

cc @trevor-scheer