Remove incorrect comment about typings being an appropriate assertion.
Adjusting the typings only works for users of TypeScript.  On the other
hand, this is a one-time assertion which happens at instantiation time which
could save a JavaScript developer from the pain of not knowing what's going
on!

There might be typing improvements to be had, but claiming it as an alternate
approach to handling this isn't correct.  The typings can still be improved,
of course.