Ensure the `executor` is set in gateway mode after failed `load` recoveries.
Previously, if the initial call to `load` failed, but its intention (fetching a
federated schema configuration in managed federation) is eventually
accomplished via a subsequent fetch (via setInterval polling), the
`executor` would not be set.

This resulted in a continued failure even if the `schema` was eventually set
since federated `schema`s require the Gateway's `executor` to do pull off
their much more complex (remote!) execution strategy!

The solution was simple since `executor` was already present on the actual
`ApolloGateway`, but that required exposing that property as a valid type to
access from the interface that `ApolloGateway` implements: `GraphQLService`.

I don't see why a `GraphQLService` wouldn't have an executor, so it seemed
appropriate to add, particularly since our only `GraphQLService` is the
`ApolloGateway` class itself.